Provide ParserStateMachine as a global pointer instead of global object
Visual C++ somehow doesn't seem to pick-up the fact that the global object is referenced and thus doesn't link in the compiled lalrc generated source file.  Using a pointer also makes the call slightly simpler in that the address of the object doesn't need to be taken when passing it to the Parser.