Use latest Forge features in Lalrc target prototype
Interpolates the portable path to the *lalrc* executable target and
stores this target as the first implicit dependency of the parse table
source file being built.

Relies on the implicit dependencies **not** being cleared in
`forge:build_visit()` as they used to be so that the implicit dependency
is still available to retrieve the path to the executable to run to
build the parse tables at build time.

Also relies on implicit conversion of target to string (the target's filename) in
forge:execute() so that the *lalrc* target can be passed directly to
forge:system() -> forge:execute().