Keep RegexSyntaxTree allocated until after states are generated
Previously the RegexSyntaxTree was a temporary that was destroyed before
states were generated.  This also destroyed the RegexNode objects that
were used in the RegexItems that define each state.  Although the states
and transitions were seemingly still correct the order in which they
were output into the generate source file was not deterministic.

Fixing this stops the generated source files from changing when they are
built again even if the source grammar hasn't changed.  It seems hard to
believe that this wasn't also causing larger problems but those don't
show up in the tests.