Check that implicit terminals only specify associativity on one symbol
Implicit terminals can be identified by their literal value or by the
identifier that names them.  But if both forms are used to specify an
associativity and precedence then it's not clear which one is correct.
Make it an error if associativity is specified for a terminal by its
literal and its identifier.