Store ParserSymbol in whitespace lexer tokens matched symbol
Previously this cast a GrammarSymbol to a ParserSymbol and then wrote
out its index -- generating a wildly wrong value.  This was okay at
runtime because the lexer only checks to see whether or not there is a
symbol and never uses its value.

However it looks like a pretty nasty error to anyone looking at the
generated source code.  Fixed by converting the GrammarSymbol to the
correct ParserSymbol when the lexer is written out.