Use correct string for lexeme when debugging parsers
This works for UTF-8 encoded strings but debugging will fail when using
UTF-16 or UTF-32 as the printf() formatted string won't deal with null
characters embedded within the wider character encodings.

Committing this for now because it doesn't break any existing
functionality but does enable parsing of wider character inputs.